requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings==2.7.1
aiofiles==24.1.0
langchain==0.3.18
PyMuPDF==1.25.3
motor==3.7.0

cohere~=5.14.0
protobuf~=5.29.3
openai~=1.63.2
anthropic~=0.49.0
mistralai~=1.5.1
pymongo~=4.11.1
pydantic~=2.10.6
langchain-core~=0.3.37
sqlalchemy~=2.0.38
langchain-community~=0.3.17
langchain-text-splitters~=0.3.6